# docs/openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Business Systems Integration Hub API",
    "version": "1.0.0",
    "description": "A comprehensive middleware platform for synchronizing data between various business systems",
    "contact": {
      "name": "API Support",
      "email": "support@integrationhub.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ],
  "paths": {
...
    "/api/file-upload/validate": {
      "post": {
        "summary": "Validate uploaded file format",
        "description": "Check if uploaded file is valid for import without processing",
        "tags": [
          "File Upload"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "File to validate"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File validation result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "valid": {
                      "type": "boolean"
                    },
                    "fileInfo": {
                      "type": "object",
                      "properties": {
                        "fileName": {
                          "type": "string"
                        },
                        "fileSize": {
                          "type": "number"
                        },
                        "mimeType": {
                          "type": "string"
                        },
                        "headers": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "totalRows": {
                          "type": "number"
                        }
                      }
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "No file uploaded or invalid file"
          }
        }
      }
    },
...
  },
...
}